requirements.txt
streamlit
pypdf2
python-docx
google-generativeai
python-dotenv